# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TimeBreaker API",
        "version": "1.0.0"
    },
    "paths": {
        "/timebreak": {
            "get": {
                "tags": [
                    "TIME BREAKDOWN"
                ],
                "summary": "Get All Time Breakdown Records",
                "description": "Get All Time Breakdown Records with filter for from_date and to_date",
                "operationId": "App\\Http\\Controllers\\TimeBreakController::index",
                "parameters": [
                    {
                        "name": "from_date",
                        "in": "query",
                        "description": "From Datetime",
                        "schema": {
                            "type": "string",
                            "example": "2020-01-01 00:00:00"
                        }
                    },
                    {
                        "name": "to_date",
                        "in": "query",
                        "description": "To Datetime",
                        "schema": {
                            "type": "string",
                            "example": "2020-02-05 12:30:00"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Time Breakdown Record Found.",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "code": 200,
                                    "data": [
                                        {
                                            "id": 1,
                                            "from_date": "2020-01-01 00:00:00",
                                            "to_date": "2020-02-05 12:30:00",
                                            "expression": [
                                                "2m",
                                                "m",
                                                "d",
                                                "2h"
                                            ],
                                            "result": {
                                                "2m": 0,
                                                "m": 1,
                                                "d": 5,
                                                "2h": 6.25
                                            },
                                            "created_at": "2021-03-05T18:19:40.000000Z",
                                            "updated_at": "2021-03-05T18:19:40.000000Z"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "TIME BREAKDOWN"
                ],
                "summary": "Create Time Breakdown",
                "description": "Create Time Breakdown from Date Range and Date Expression",
                "operationId": "App\\Http\\Controllers\\TimeBreakController::breakTime",
                "requestBody": {
                    "description": "JSON Payload",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "from_date": {
                                        "description": "From Datetime",
                                        "type": "string",
                                        "example": "2020-01-01 00:00:00"
                                    },
                                    "to_date": {
                                        "description": "To Datetime",
                                        "type": "string",
                                        "example": "2020-02-05 12:30:00"
                                    },
                                    "expression": {
                                        "description": "An array of expression on how to breakdown the date range",
                                        "type": "array",
                                        "items": {
                                            "type": "array",
                                            "items": {}
                                        },
                                        "example": [
                                            "2m",
                                            "m",
                                            "d",
                                            "2h"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "IP Address creation successful",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": [
                                    {
                                        "success": true,
                                        "code": 200,
                                        "data": {
                                            "2m": 0,
                                            "m": 1,
                                            "d": 5,
                                            "2h": 6.25
                                        }
                                    }
                                ]
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid given data",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": [
                                    {
                                        "success": false,
                                        "code": 422,
                                        "data": {
                                            "message": "The given data was invalid.",
                                            "details": {
                                                "from_date": [
                                                    "The from date field is required."
                                                ],
                                                "to_date": [
                                                    "The to date field is required."
                                                ],
                                                "expression": [
                                                    "The expression field is required."
                                                ]
                                            }
                                        }
                                    }
                                ]
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    }
}